YAML CPP: Move function into anonymous namespace
