+++ src/plugins/yamlcpp/log.hpp
@@ -14,7 +14,7 @@
 
 #ifdef HAVE_LOGGER
 
-namespace yamlcpp
+namespace
 {
 
 /**
@@ -22,9 +22,24 @@ namespace yamlcpp
  *
  * @param keys This parameter stores the key set this function prints.
  */
-void logKeySet (kdb::KeySet const & keys);
+void logKeySet (kdb::KeySet const & keys)
+{
+	for (auto key : keys)
+	{
+		std::string metadata;
+		key.rewindMeta ();
+		while (kdb::Key meta = key.nextMeta ())
+		{
+			metadata += ", “" + meta.getName () + "”: “" + meta.getString () + "”";
+		}
 
+		ELEKTRA_LOG_DEBUG ("\t“%s”: “%s”%s", key.getName ().c_str (),
+				   key.getBinarySize () == 0 ? "NULL" : key.isBinary () ? "binary value!" : key.getString ().c_str (),
+				   metadata.c_str ());
+	}
 }
+
+} // end namespace
 
 #endif // HAVE_LOGGER
